requirements.txt
opencv-python
numpy
Pillow
pyautogui